+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/ui/home/explore/Navigation.kt
@@ -4,7 +4,6 @@ import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.navigation
-import indi.dmzz_yyhyy.lightnovelreader.ui.home.explore.expanded.exploreExpandDestination
 import indi.dmzz_yyhyy.lightnovelreader.ui.home.explore.home.exploreHomeDestination
 import indi.dmzz_yyhyy.lightnovelreader.ui.home.explore.search.exploreSearchDestination
 import io.nightfish.lightnovelreader.api.Route
@@ -15,7 +14,6 @@ fun NavGraphBuilder.exploreNavigation() {
         startDestination = Route.Main.Explore.Home,
     ) {
         exploreHomeDestination()
-        exploreExpandDestination()
         exploreSearchDestination()
     }
 }